mise en place d'une structure de style plus maintenable.
