+++ app/src/components/Publication/CaptionInput.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { View, TextInput, Image } from 'react-native';
-import styles from '../../styles/publicationStyles';
+import styles from '../../styles/captionInputStyles';
 
 interface CaptionInputProps {
   value: string;
