+++ app/src/components/UserProfile.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
-import { View, Text, StyleSheet, Image, TouchableOpacity, Alert } from 'react-native';
+import { View, Text, Image, TouchableOpacity, Alert } from 'react-native';
 import FontAwesome from 'react-native-vector-icons/FontAwesome';
 import { useAuth } from '../context/AuthContext';
+import styles from '../styles/userProfileStyles';
 
 interface UserProfileProps {
   navigateToHome?: () => void;
@@ -35,7 +36,7 @@ const UserProfile: React.FC<UserProfileProps> = ({ navigateToHome }) => {
       <View style={styles.header}>
         <Image 
           source={{ 
-            uri: user?.avatar || 'https://cdn.builder.io/api/v1/image/assets/TEMP/f3d67917c6442fafa158ab0bdb706f7194e5a329b13b72692715ec90abbf8ce7?placeholderIfAbsent=true'
+            uri: user?.photoURL || 'https://cdn.builder.io/api/v1/image/assets/TEMP/f3d67917c6442fafa158ab0bdb706f7194e5a329b13b72692715ec90abbf8ce7?placeholderIfAbsent=true'
           }} 
           style={styles.profileImage} 
         />
@@ -72,57 +73,5 @@ const UserProfile: React.FC<UserProfileProps> = ({ navigateToHome }) => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#FFF',
-  },
-  header: {
-    alignItems: 'center',
-    paddingVertical: 30,
-    borderBottomWidth: 1,
-    borderBottomColor: '#E8ECF4',
-  },
-  profileImage: {
-    width: 100,
-    height: 100,
-    borderRadius: 50,
-    marginBottom: 15,
-  },
-  username: {
-    fontSize: 24,
-    fontWeight: 'bold',
-    marginBottom: 5,
-    color: '#1E232C',
-  },
-  email: {
-    fontSize: 16,
-    color: '#6A707C',
-    marginBottom: 10,
-  },
-  infoContainer: {
-    padding: 20,
-  },
-  infoItem: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    paddingVertical: 15,
-    borderBottomWidth: 1,
-    borderBottomColor: '#E8ECF4',
-  },
-  infoText: {
-    fontSize: 16,
-    marginLeft: 15,
-    color: '#1E232C',
-  },
-  logoutItem: {
-    marginTop: 20,
-    borderBottomWidth: 0,
-  },
-  logoutText: {
-    color: '#FF4C4C',
-  },
-});
 
 export default UserProfile;
