+++ app/src/screens/LoadingScreen.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import { View, ActivityIndicator, StyleSheet, Text } from 'react-native';
+import { View, ActivityIndicator, Text } from 'react-native';
+import styles from '../styles/loadingStyles';
 
 const LoadingScreen: React.FC = () => {
   return (
@@ -9,19 +10,5 @@ const LoadingScreen: React.FC = () => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#FFFFFF',
-  },
-  text: {
-    marginTop: 10,
-    fontSize: 16,
-    color: '#1E232C',
-  },
-});
 
 export default LoadingScreen;
