+++ app/src/styles/bottomNavStyles.ts
@@ -1,0 +1,28 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flexDirection: "row",
+    backgroundColor: "#fff",
+    paddingVertical: 8,
+    borderTopWidth: 1,
+    borderTopColor: "rgba(0, 0, 0, 0.1)",
+  },
+  tab: {
+    flex: 1,
+    alignItems: "center",
+    justifyContent: "center",
+    paddingVertical: 8,
+  },
+  tabText: {
+    fontSize: 12,
+    marginTop: 4,
+    color: "#6A707C",
+  },
+  activeTabText: {
+    color: "#000",
+    fontWeight: "bold",
+  },
+});
+
+export default styles;
