+++ app/src/styles/createJobOfferStyles.ts
@@ -1,0 +1,77 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: "rgba(251, 249, 250, 1)",
+  },
+  topBar: {
+    backgroundColor: "#fff",
+    shadowColor: "#000",
+    shadowOffset: {
+      width: 0,
+      height: 2,
+    },
+    shadowOpacity: 0.12,
+    shadowRadius: 3,
+    elevation: 4,
+    minHeight: 60,
+  },
+  topBarContent: {
+    flexDirection: "row",
+    alignItems: "center",
+    paddingHorizontal: 16,
+    paddingVertical: 12,
+    gap: 12,
+  },
+  topBarTitle: {
+    fontSize: 20,
+    fontWeight: "500",
+    color: "#000",
+  },
+  content: {
+    flex: 1,
+    padding: 12,
+  },
+  inputContainer: {
+    marginBottom: 20,
+  },
+  inputLabel: {
+    fontSize: 16,
+    fontWeight: "500",
+    color: "#000",
+    marginBottom: 4,
+  },
+  textInput: {
+    borderWidth: 1,
+    borderColor: "rgba(0, 0, 0, 0.1)",
+    borderRadius: 6,
+    padding: 12,
+    fontSize: 16,
+    color: "#000",
+    backgroundColor: "#fff",
+  },
+  textAreaInput: {
+    height: 100,
+    textAlignVertical: "top",
+  },
+  inputInfo: {
+    fontSize: 14,
+    color: "rgba(0, 0, 0, 0.5)",
+    marginTop: 4,
+  },
+  submitButton: {
+    backgroundColor: "#000",
+    borderRadius: 8,
+    padding: 12,
+    alignItems: "center",
+    marginVertical: 20,
+  },
+  submitButtonText: {
+    color: "#fff",
+    fontSize: 16,
+    fontWeight: "500",
+  },
+});
+
+export default styles;
