+++ app/src/styles/loadingStyles.ts
@@ -1,0 +1,17 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '#FFFFFF',
+  },
+  text: {
+    marginTop: 10,
+    fontSize: 16,
+    color: '#1E232C',
+  },
+});
+
+export default styles;
