+++ app/src/styles/profileComponentStyles.ts
@@ -1,0 +1,9 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  spacerView: {
+    width: 24
+  }
+});
+
+export default styles;
