+++ app/src/styles/registerComponentStyles.ts
@@ -1,0 +1,9 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  flexContainer: {
+    flex: 1
+  }
+});
+
+export default styles;
