+++ app/src/styles/userProfileStyles.ts
@@ -1,0 +1,55 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#FFF',
+  },
+  header: {
+    alignItems: 'center',
+    paddingVertical: 30,
+    borderBottomWidth: 1,
+    borderBottomColor: '#E8ECF4',
+  },
+  profileImage: {
+    width: 100,
+    height: 100,
+    borderRadius: 50,
+    marginBottom: 15,
+  },
+  username: {
+    fontSize: 24,
+    fontWeight: 'bold',
+    marginBottom: 5,
+    color: '#1E232C',
+  },
+  email: {
+    fontSize: 16,
+    color: '#6A707C',
+    marginBottom: 10,
+  },
+  infoContainer: {
+    padding: 20,
+  },
+  infoItem: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    paddingVertical: 15,
+    borderBottomWidth: 1,
+    borderBottomColor: '#E8ECF4',
+  },
+  infoText: {
+    fontSize: 16,
+    marginLeft: 15,
+    color: '#1E232C',
+  },
+  logoutItem: {
+    marginTop: 20,
+    borderBottomWidth: 0,
+  },
+  logoutText: {
+    color: '#FF4C4C',
+  },
+});
+
+export default styles;
